Move ws and branchID resolve function to utils directory
