+++ plugins/dagsvisualizer/frontend/src/components/MessageInfo.tsx
@@ -3,7 +3,7 @@ import Card from 'react-bootstrap/Card';
 import ListGroup from 'react-bootstrap/ListGroup';
 import { inject, observer } from 'mobx-react';
 import TangleStore from 'stores/TangleStore';
-import { resolveBase58BranchID } from 'utils';
+import { resolveBase58BranchID } from 'utils/BranchIDResolver';
 import * as dateformat from 'dateformat';
 
 interface Props {
