+++ plugins/dagsvisualizer/frontend/src/utils/WS.ts
